requirements.txt
annotated-types==0.7.0
anthropic==0.39.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
